Move collection code into separate file
